third_party/blink/web_tests/platform/generic/fast/forms/file/input-file-label-expected.txt

  

This tests the label of a file chooser button.

PASS The label of a single file chooser button is "Choose File".
PASS The label of a multiple file chooser button is "Choose Files".
PASS Initially, the label of a file chooser button is "Choose File".
PASS Set "multiple" attribute, then the label of the file chooser button becomes "Choose Files".
PASS Unset "multiple" attribute, then the label of the file chooser button becomes "Choose File".
PASS successfullyParsed is true

TEST COMPLETE
